Advances on parser
It seems that the deserializer is ready.

It allows to parse a file in one single step (although not recommendable), and also obtain a single unit of sensors data.
Check the SensingUnit and SensorDataBlock classes.
The idea is that the SensorDataBlock concentrates information about a single sensing intervention (for instance, one per GPS, multiple for accelerometer).
The SensingUnit concentrates a list of SensorDataBlocks and at the end, a sensing file could be represented by a list of SensingUnit.